Add files via upload
It is an analysis to identify the root cause of the problem (i.e. cancellation and non-availability of cars) and recommend ways to improve the situation. As a result of the analysis, we were able to present to the client the root cause(s) and possible hypotheses of the problem(s) and recommend ways to improve them.